[MOVE] #30 - presentation 패키지 구조화
