+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.MainActivity">
+        tools:context=".presentation.ui.main.MainActivity">
 
         <TextView
             android:layout_width="wrap_content"
